Create project from template fixes
