+++ src/com/butent/bee/client/modules/projects/ProjectsGrid.java
@@ -46,27 +46,6 @@ class ProjectsGrid extends AllProjectsGrid {
     return true;
   }
 
-  @Override
-  public boolean beforeAction(Action action, GridPresenter presenter) {
-
-    if (Action.ADD.equals(action) && BeeKeeper.getUser().isDataVisible(VIEW_PROJECT_TEMPLATES)) {
-      FormFactory.createFormView(ProjectConstants.FORM_NEW_PROJECT_FROM_TASK, null, null, false,
-          getNewProjectFormInterceptor(),
-          new FormFactory.FormViewCallback() {
-            @Override
-            public void onSuccess(FormDescription formDescription, FormView form) {
-              if (form != null) {
-                form.start(null);
-                Global.showModalWidget(form.getCaption(), form.asWidget());
-              }
-            }
-          });
-      return false;
-    }
-
-    return true;
-  }
-
   private static Filter getGridFilter() {
     Filter isOwner = Filter.equals(COL_PROJECT_OWNER, USER_ID);
     Filter isProjectUser =
@@ -74,95 +53,4 @@ class ProjectsGrid extends AllProjectsGrid {
             AdministrationConstants.COL_USER, USER_ID));
     return Filter.or(isOwner, isProjectUser);
   }
-
-  private AbstractFormInterceptor getNewProjectFormInterceptor() {
-    return new AbstractFormInterceptor() {
-
-      private static final String NAME_PROJECT_TEMPLATE = "ProjectTemplate";
-      private static final String NAME_CREATE = "Create";
-      private static final String NAME_CANCEL = "Cancel";
-
-      private UnboundSelector templateSelector;
-      private Button createButton;
-      private Button cancelButton;
-
-      @Override
-      public void afterCreateWidget(String name, IdentifiableWidget widget,
-          FormFactory.WidgetDescriptionCallback callback) {
-
-        switch (name) {
-          case NAME_PROJECT_TEMPLATE:
-            templateSelector = widget instanceof UnboundSelector ? (UnboundSelector) widget : null;
-            break;
-          case NAME_CREATE:
-            createButton = widget instanceof Button ? (Button) widget : null;
-
-            if (createButton == null) {
-              break;
-            }
-
-            createButton.addClickHandler(new ClickHandler() {
-              @Override
-              public void onClick(ClickEvent event) {
-                Popup.getActivePopup().close();
-
-                if (templateSelector == null) {
-                  createSimpleProject();
-                } else if (BeeUtils.isEmpty(templateSelector.getValue())) {
-                  createSimpleProject();
-                } else {
-                  createProjectFromTemplate(templateSelector.getRelatedRow());
-                }
-              }
-            });
-            break;
-          case NAME_CANCEL:
-            cancelButton = widget instanceof Button ? (Button) widget : null;
-
-            if (cancelButton == null) {
-              break;
-            }
-
-            cancelButton.addClickHandler(new ClickHandler() {
-              @Override
-              public void onClick(ClickEvent event) {
-                if (templateSelector != null) {
-                  templateSelector.setValue(null, true);
-                }
-                Popup.getActivePopup().close();
-              }
-            });
-            break;
-          default:
-            super.afterCreateWidget(name, widget, callback);
-        }
-
-      }
-
-      @Override
-      public FormInterceptor getInstance() {
-        return this;
-      }
-
-    };
-  }
-
-  private static void createSimpleProject() {
-    RowFactory.createRow(VIEW_PROJECTS);
-  }
-
-  private void createProjectFromTemplate(final IsRow templateRow) {
-    ProjectsKeeper.createProjectFromTemplate(templateRow, new RowCallback() {
-      @Override
-      public void onSuccess(final BeeRow projectRow) {
-
-        if (getGridView() != null) {
-          getGridView().notifyInfo(Localized.getMessages()
-              .newProjectCreated(projectRow.getId()));
-        }
-
-      }
-    });
-  }
-
 }
